sched: ems: Separate the 'high capacity' hint from 'performance-preferred'
We're encountering a lot of heavy UI stuttering if we place top-app tasks with the 'prefer-perf' hint as preemption is always avoided on CPUs where these tasks currently reside. The problem is, we don't want 'top-app' to be only using the 'prefer-idle' hint as we want to put these tasks onto performant cores as much as possible. I've also found out that using the 'service' path for all top-app tasks seem to cause UX stutters from time-to-time and we may want to use the 'ontime' path for some tasks.

Introduce a new 'prefer-high-cap' hint, inspired from Google's changes on the Pixel, instead for checking when the task either wants a high-capacity or a performant CPU. This way, we can still take advantage of the PCF for top-app, which is a lower priority over 'ontime wakeup'. This way, we can keep performance for all top-app tasks without the worry of placing them in the 'service' path.

Also move the kernel's own 'prefer-perf' hint to this, as this is triggered on device wake, interaction, and app launch. This is also what Google uses the 'prefer-high-cap' hint for[1].

[1]: https://android.googlesource.com/device/google/redbull/+/refs/heads/android12L-s1-release/powerhint.json

Signed-off-by: John Vincent <git@tenseventyseven.cf>