sched: ems: energy: Use EMS' 5.10 implementation of CPU 'performance' checking
Something I've noticed while trying to revamp the 'band' placement path is that is_slowest_cpu() returns true on CPU4, which is a higher-capacity CPU as that is part of the big cluster. This can be (very) problematic as we cannot guarantee the accuracy of this, and can cause (tons of) issues as this is used on multiple scheduler placement paths.

I've found that Exynos' 5.10 kernels (more specifically, the Exynos 1280) have a very similar API to this, which from testing is more accurate than what we have right now. Get this new API and make it work for our platform (which is very easy, btw).

Signed-off-by: John Vincent <git@tenseventyseven.cf>